move decorator to better place
