+++ bokeh/server/protocol/messages/__init__.py
@@ -14,6 +14,24 @@ def register(cls):
     index[key] = cls
     return cls
 
+def _add_buffer(self, buf_header, buf_payload):
+    raise ProtocolError("")
+
+def _write_buffers(self, conn):
+    return 0
+
+def _is_complete(self):
+    return True
+
+def nobuffers(cls):
+    '''
+
+    '''
+    cls.add_buffer = _add_buffer
+    cls.write_buffers = _write_buffers
+    cls.is_complete = _is_complete
+    return cls
+
 from .ack import *
 from .ok import *
 from .error import *
